heavy refactor to encapsulate template walking.  All that is exposed is verifiers
