+++ pkg/kudoctl/cmd/package_verify.go
@@ -6,13 +6,14 @@ import (
 
 	"github.com/kudobuilder/kudo/pkg/kudoctl/cmd/verify"
 	"github.com/kudobuilder/kudo/pkg/kudoctl/packages/reader"
+	"github.com/kudobuilder/kudo/pkg/kudoctl/packages/verifier"
 
 	"github.com/gosuri/uitable"
 	"github.com/spf13/afero"
 	"github.com/spf13/cobra"
 )
 
-//TODO (kensipe): add long desc
+// package verify provides verification or linting checks against the package passed to the command.
 
 type packageVerifyCmd struct {
 	fs  afero.Fs
@@ -59,7 +60,7 @@ func verifyPackage(fs afero.Fs, path string, out io.Writer) error {
 	return fmt.Errorf("package verification errors: %v", len(errors))
 }
 
-func printErrors(out io.Writer, errors verify.ParamErrors) {
+func printErrors(out io.Writer, errors verifier.ParamErrors) {
 	table := uitable.New()
 	table.AddRow("Errors")
 	for _, err := range errors {
@@ -68,7 +69,7 @@ func printErrors(out io.Writer, errors verify.ParamErrors) {
 	fmt.Fprintln(out, table)
 }
 
-func printWarnings(out io.Writer, warnings verify.ParamWarnings) {
+func printWarnings(out io.Writer, warnings verifier.ParamWarnings) {
 	table := uitable.New()
 	table.AddRow("Warnings")
 	for _, warning := range warnings {
